Use bitset package

+++ ecs/component.go
@@ -1,7 +1,7 @@
 package ecs
 
 // MaxEntities is the maximum number of components that can be created per entity.
-const MaxComponents uint8 = 32
+const MaxComponents uint = 32
 
 // ComponentType is a unique identifier for a type of component.
 type ComponentType uint8
